Restore Kelvin homoglyph for `K`
The issue behind #2 is that the Unicode canonical case-folding document
that vim uses to build its case-folding tables folds `\u212a` to `k`.
This plugin should be ignoring case in its matches, anyway, to be more
accurate, so turn case matching on explicitly instead of relying on
other syntax files doing it for us.